More renaming and warning fixes
